generator refactor
